backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
httpx
Pillow
transformers
torch
